Housekeeping - Switch to objection for query
The original query was built when we thought we needed to get billing account details as part of the query. This was because we thought we'd be matching as we iterated the charge versions we fetch and their charge elements.

The version of the supplementary billing engine that finally switched had opted to pre-generate billing data and then iterate it. The need to access billing account information on the transaction had become redundant.

We should have noticed at the time but it's never too late!

Because we can drop those elements we can switch to using Objection which tidies up the query some, for example, the `purposes` field now comes through as proper JSON.

These changes to the what `FetchPreviousTransactionsService` returns also means we can remove some of the logic in `ReverseTransactionsService`.