Supplementary billing housekeeping (#857)
https://eaflood.atlassian.net/browse/WATER-4403

> Part of changes to exclude the current financial year from supplementary billing if the annual bill run has not been completed

We have recently migrated the SROC annual billing engine into our project ( [WATER-4345](https://eaflood.atlassian.net/browse/WATER-4345) ). The primary drive to do this was to enable our users to run annual bill runs in April at the same time as licensees are submitting their return data. This is when the service is under its greatest load so we wanted to ensure annual billing could take advantage of the improvements our supplementary billing engine has over the legacy one.

But having built the annual billing engine, we found there were improvements we could make to our supplementary billing engine!

So, we're doing this in two parts. This first part covers some housekeeping changes we've spotted that could be made.

---

- `ProcessBillRunService` should have used `currentTimeInNanoseconds` to get the start time
- Test for `HandleErroredBillRunService` was in the wrong place
- Stop passing objects around when all we need are ID's
- Update `FetchPreviousTransactionsService` to use Objection rather than Knex raw
- Drop properties we're not using from `FetchPreviousTransactionsService` query (which means we can simplify the conversion in `ReverseTransactionsService`)