Fix for view licence abstraction conditions V2 (#853)
https://eaflood.atlassian.net/browse/WATER-4332

> @Cruikshanks stab at the problem!

As part of the testing for the new view licence summary page, QA pointed out that the more complicated licences did not match the legacy system. We would see different results for the 'Abstraction conditions' section.

The first thing we'd overlooked was the need to ensure we were only looking at the 'current' licence version. After more investigation, we then discovered that the permit licence JSON blob the legacy code is using is replicated in tables that already exist in the `water` schema. We found we could access the same data using these tables to get a matching result to the legacy view.

This change adds a new service to fetch the abstraction data we need and updates the presenter to generate the distinct list of conditions to display.