Refactor FetchBillingAccountsService
The primary enabler in the annual billing improvements was moving to fetching the data at the billing account level. Currently, in supplementary we are fetching the charge versions and then pre-grouping them into bills (the pre-generate process). Part of that process was fetching the billing account details for each charge version group.

If we start at the billing account level, we can remove that complexity _and_ more easily batch the processing of bills.

So, the first step is to refactor `FetchBillingAccountsService` to get _everything_ needed just as we do in annual billing.