Refactor ProcessBillRunService
Now we've replaced FetchChargeVersionsService with FetchBillingAccountsService we need to update ProcessBillRunService to use it.

As well as that we make some other improvements.

- put functions into alphabetical order
- removed unnecessary _logLevel() function
- update comments
- remove `is` prefix from boolean

We also look ahead to the refactoring of `ProcessBillingPeriodService`. Currently, it is returning a flag to say whether any bills were generated as well as what licence ID's were processed. As it is iterating the charge versions it made sense at the time.

Now, we're bringing ProcessBillingPeriodService inline with what we are doing in the annual engine. Plus, it's easier to follow if we figure this out within the ProcessBillRunService at the same time as we get fetch the billing accounts for the period.