Update ProcessBillingPeriodService
Obviously, the key switch is we are iterating billing accounts. This also means we can bring in the same batching process for bills as we use in the annual billing engine. The PreGenerateBillingDataService is removed because we don't need to group charge versions by billing account and fetch the billing account details separately.

Other than that we do general clean up such as ensuring the functions are alphabetical etc.