Fix determine billing periods (#859)
https://eaflood.atlassian.net/browse/WATER-4375

To implement changes to support new two-part tariff bill run options we needed to update the create bill run journey. We took advantage of this to migrate the functionality into [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system).

In our new journey to determine if there are any live bill runs that would block a new bill run from proceeding, we have to determine the financial end year for the proposed bill run. We do this in `app/services/bill-runs/setup/exists.service.js`.

Having determined the year we pass this into our existing `StartBillRunProcessService` for consistency.

What we had overlooked was that if the year is set it will then pass that on to `DetermineBillingPeriodsService` which, if provided with a financial year will _only_ generate a billing period for that year. This breaks supplementary billing which for SROC needs multiple billing periods to be determined.

We'll be revisiting `DetermineBillingPeriodsService` soon as part of our work to [Exclude Financial Year from Supplementary Billing where Annual Bill has not been run for region & year](https://eaflood.atlassian.net/browse/WATER-4403). But for now, we just need to get supplementary billing working again.