Fix legacy requests for /heath/info page
https://eaflood.atlassian.net/browse/WATER-4375

Recently we extended the functionality of our code that manages sending requests to the legacy services to [Add defra-user-id header to legacy requests](https://github.com/DEFRA/water-abstraction-system/pull/796). This is in preparation of us managing the setup bill run process as part of changes we're doing to support two-part tariff. Some of the legacy endpoints expect the header and use it to authorize the request being made.

When we did this we overlooked the requests being made in `app/services/health/info.service.js`. The calls it was making told our logic to send the requests to the base domains for each legacy service, for example, `http://localhost:8001`. But because of the new param and a change in their order, this was being received as the user ID arg. So, we broke a number of the requests. 🤦

This change fixes the requests being made in `InfoService`.