Fix issue with duplicated abstraction points for return requirements points page
https://eaflood.atlassian.net/browse/WATER-4295

https://github.com/DEFRA/water-abstraction-system/pull/833

This was a ticket that was thought to have been completed, but upon QA testing by Ven, some licences seem to have
duplicated 'abstraction points'. There was a previous check within the code preventing this but it was removed as no
duplicating 'point_details' were encountered at that time.

This PR will be focused on preventing duplicating abstraction points being present within the list of checkbox points on
the return requirements points page.