Add failing scenario for ProcessBillingPeriodService
When running the refactored engine through our acceptance tests we had one that repeatedly failed. Diving in we found that our refactoring had introduced a flaw.

Because we used the updated process we'd adopted for the annual bill run engine we copied across where determining the charge period was happening.

This was before we generate and process transactions for the charge version. It could be that a billing account was replaced by another prior to the billing period being processed. This would mean it's end date would be earlier than the billing period's start date resulting in an invalid charge period.

For annual that means we can disregard it completely. For supplementary that just means we shouldn't bother to generate a transaction. It doesn't mean we shouldn't process the charge version and look for previous transactions which might result in a credit being raised.

It means adding an involved and complex unit test. But it's a scenario we can't afford to get wrong.