Fix bill run exists page for bill runs in review (#863)
https://eaflood.atlassian.net/browse/WATER-4375

We recently migrated the setup bill run journey into this project which meant we also needed to migrate the page which displays when an existing bill run is found.

In testing, we have spotted an issue with the page. If the match is a two-part tariff bill run in review, the 'Go to bill run' link takes them to the view bill run page. This means the user just sees an empty bill run when they should be seeing the appropriate review page.

This change fixes the issue.