Amend the billing period in the tests
We want to test that an SROC charge version that ends before the billing starts is excluded. But the billing period being used in the tests is for the first SROC billing period. We can't replicate an SROC charge version that starts before 2022-04-01.

So, we move the billing period forward a year to give us an SROC year before hand we can use.