Remove redundant fetch for region code in service
https://eaflood.atlassian.net/browse/WATER-4375

Spotted whilst working on [Exclude ended charge versions from 2PT billing](https://github.com/DEFRA/water-abstraction-system/pull/865) that we pass in the region ID to `FetchChargeVersionsService` but then do an additional query to identify the NALD region code for that region.

We then use the NALD region code to filter the charge versions when we could just use the region ID directly.

This change removes the redundant fetch.