Fix licence issues failing to insert for review
https://eaflood.atlassian.net/browse/WATER-4375

In recent testing of triggering an SROC two-part tariff bill run, we've seen errors being thrown by the `PersistAllocatedLicenceToResultsService`.

After investigating it is not the issue. It was being asked to persist every issue for every element and return found by `DetermineLicenceIssuesService` in its `issues` field.

No de-duping of the issues is taking place so its blowing past the 255 char limit for the DB field for some licences.

This then causes the bill run to fail and the system app to crash, but we'll deal with that in another fix.

This change will de-dupe the issues found _before_ they are persisted in order to keep under the 255 char limit.