Add migration to alter issue column types
As pointed out in the review, even with the duplicates removed if a licence had every issue we'd be trying to set `issues` as `'Abstraction outside period, Aggregate, Checking query, No returns received, Over abstraction, Overlap of charge dates, Returns received but not processed, Returns received late, Return split over charge references, Some returns not received, Unable to match return'`. This is more than 255 characters so the error would arise again.

So, if we alter the column types to be `text` there is no limit (in practical terms) to what we can insert. Just to be on the safe side we make the same change to the returns and charge elements tables as well.