Fix FetchMatchingBillRunService for PRESROC 2PT (#870)
https://eaflood.atlassian.net/browse/WATER-4375

In testing, it was found that when you complete the journey to set up a PRESROC two-part tariff bill run and select 'winter and all year' as the season you get a match for the summer 2PT bill run in that year returned.

After investigating the issue it's not that we are matching the wrong season but that we are not filtering by season at all. If in the environment where the tests were being run the winter 2PT had been created last then it would always be the result returned first.

The flaw is a result of initially passing in the values selected in the UI raw, and then switching to transformed values. We always get values back from the UI as strings and this is what we are matching financial year against to determine if we need to apply the season clause to our query. But at some point in our development, we changed that to a number, which means it now never matches.

```javascript
  // financialYearEnding is 2022 not '2022'!
  if (['2022', '2021'].includes(financialYearEnding)) {
    query.where('summer', summer)
  }
```

This change fixes the issue.