Fix licence issues failing to insert for review (#867)
https://eaflood.atlassian.net/browse/WATER-4375

In recent testing of triggering an SROC two-part tariff bill run, we've seen errors thrown by the `PersistAllocatedLicenceToResultsService`.

After investigating it is not the issue. It was being asked to persist every issue for every element and return found by `DetermineLicenceIssuesService` in its `issues` field.

No de-duping of the issues is taking place so it's blowing past the 255-char limit for the DB field for some licences.

This causes the bill run to fail and the system app to crash (but we'll deal with that in another fix!)

This change will de-dupe the issues found _before_ they are persisted to keep the entry to a minimum. But should a licence have every single issue the resulting value would still be more than 255 characters. 

So, we also add a migration to alter the `issues` column in `review_licences`, `review_charge_elements` and `review_returns` from `varchar(255)` to `text` which has no limit (in practical terms).