Refactor the service
It now expects the bill run type and uses this information to decide how many billing periods to return and not the financial year.

This sets us up to be able to tell the service we need the billing periods for a supplementary bill run, but start in year X instead of the current financial year.