Add a remove duplicate licence feature
https://eaflood.atlassian.net/browse/WATER-4429

Usually once or twice a month we have to deal with a duplicate licence in the service. It is not actually a duplicate but an issue with NALD (the upstream service) that our [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) does not handle.

We believe when a change is made to a licence in NALD it requires users to enter the licence reference again. This creates a new record in NALD but our import service recognises it is for the same licence because the references match.

However, NALD is known for poor validation and doesn't trim data entered. So, often a user will accidentally add an extra space or newline character when entering the licence reference again. On import, computers being the way they are, will see this as a new licence so will create a new licence record with the same 'bad' reference.

This causes an error for anything that attempts to retrieve the licence by reference in the service. The most visible casualty is the search page.

We _should_ be fixing the import. But we're only just diving into that black-box of code and are reticent to change things we don't understand. Plus the reason we are now looking at the import code is because of the strong likelihood we will have to build a new import service this year.

So, to avoid constantly having to create migration scripts and to allow us to respond sooner when an issue is spotted, this change adds a tool that will allow us to delete any 'bad' licence records.