Add our new routes to /data
We'll need a page that allows someone to enter the reference to be deleted. We'll not be adding a link to it anywhere as it's only us on the dev team that we intend to be using it. So you'll need to type it in when you need it!

Because of that we've added it to our `/data` root. It is related to licences but we think as it is intended for our use `/data` is a better home.

We've scoped it to `billing` which is a role a number of us have. Billing & Data is also the team that notices these problem licences and deal with them in NALD. So, with there being no 'admin' or 'dev' scope we think this is the best available.