Fix send bill run unflagging licences (#884)
https://eaflood.atlassian.net/browse/WATER-4390

In a recent change we [Migrated the legacy send bill run functionality](https://github.com/DEFRA/water-abstraction-system/pull/771) to our service.

We thought all was well but then recently spotted an issue. When we are 'sending' a bill run (telling the CHA to generate the transaction file for SOP that results in the bills being sent to customers) we are calling our `SubmitSendBillRunService` which in turn is calling `UnflagBilledLicencesService`.

Only it shouldn't be doing this _all_ the time. It should only be calling that service if the bill run is supplementary.

This change fixes the issue.