Mark Progress when reviewing Licence in 2PT Bill Run
https://eaflood.atlassian.net/browse/WATER-4215

Add a button to the review individual licence screen called “Mark Progress”. Clicking this button will mark the licence as in progress, changing the button to “remove progress from licence” and adding a banner to the top of the page. If you then go to the review licences page, that licence will have a tick against it in the progress column.