Fix review licence page
https://eaflood.atlassian.net/browse/WATER-4219

Following from the testing of the above ticket, two issues with the page was discovered. The first being where we are meant to display the charge reference billable volume this is actually displaying the charge elements and the second issue being a piece of text missing at the top of the page. This PR is to fix those issues.