New job to add new & updated licences to workflow
https://eaflood.atlassian.net/browse/WATER-4437

**Context**

The legacy [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) has the scheduled job [Licence not in charge version workflow](https://github.com/DEFRA/water-abstraction-team/blob/main/jobs/service.md#licence-not-in-charge-version-workflow). It's intent is to add a `water.charge_version_workflow` record any time it sees a new or updated licence.

Our understanding is this is intended to keep the licence from being included in a bill run until Billing & Data have had a chance to review its charge versions.

We have had a call logged with us because this the job doesn't appear to be running any more. Billing & Data know they have made changes in the upstream service NALD which should have resulted in licences being added to workflow but hasn't.

Our digging into the logs has found this job runs sporadically at best! There have been been periods spanning weeks where the job has not run, even though it is scheduled to run 4 times a day! We believe the fact it is annual billing time has has led to a renewed focus and so the team have spotted something is not right.

Why the legacy job is being sporadic we cannot answer. The job is running using [BullMQ](https://docs.bullmq.io/) message queue architecture. In this case the schedule is set using config passed to BullMQ when creating the [repeatable job](https://docs.bullmq.io/guide/jobs/repeatable). Everything appears to be correct. But we've seen this behaviour before and we know we are now 4 versions behind the current version of BullMQ.

Not wishing to plough more time and resource into confirming exactly why BullMQ is being flaky we're are instead using this as an opportunity to migrate the job away from the legacy code.

**Change**

This adds a new endpoint to `/jobs` that will allow us to trigger the job either manually or using a schedule Jenkins job (which so far has proven to be 100% reliable 😁💪).

The legacy job, (once we figured out what was just noise!) simply looks for licence versions that were created in the last 2 months that don't have a charge version workflow record. A tweak was made to the query to exclude licences that are linked to in-progress bill runs as this caused errors for the legacy billing engine (see WATER-3528).

Our job does the same though we've cleaned up the query that identifies the licences versions that need adding. Most importantly, there is not a queue in sight!