Returns required journey - Notes page
https://eaflood.atlassian.net/browse/WATER-4269

This branch will add a notes feature to add a note to the return requirement.

The page will have a text entry box for notes and a character count below it to display how many characters have been entered.
Validation will ensure no more than 500 characters are entered.

The content will be saved in the session until the user submits the application.