Update /jobs/export to be a POST request
https://eaflood.atlassian.net/browse/WATER-3962

Jobs that we trigger we always use a POST request. It just helps ensure that a request was intentional and not from something trawling the site and making an inadvertent `GET` request.

We overlooked this when we added `/jobs/export` and were reminded of it again when working on [New job to add new & updated licences to workflow](https://github.com/DEFRA/water-abstraction-system/pull/903).

So, this is a bit of housekeeping to get the job inline with the others.