Fix and update time-limited job
https://eaflood.atlassian.net/browse/WATER-3486

In [Remove licences with charge elements with approaching time limits from billing](https://github.com/DEFRA/water-abstraction-system/pull/443) we added a background job to do just that. Then in [Refactor to use new view based WATER models](https://github.com/DEFRA/water-abstraction-system/pull/569) we updated how we interact with the DB.

We updated the service which fetches the data for the job. But we didn't clock that this results in the job erroring because we are no longer passing in `licenceId`. Without it we error trying to create the `workflow` record in the DB because of a constraint on the field.

We spotted this when working on [New job to add new & updated licences to workflow](https://github.com/DEFRA/water-abstraction-system/pull/903) so this change is mainly about fixing the bug.

But we also realised one of the reasons we'd failed to spot the issue was a lack of logging. So, we also updated the job to use our pattern of logging how long the process took. We also take an idea from the new job and set a flag on the `workflow` record so we can spot _why_ it was created.