Update fetch service to return charge version ID
Taking a lesson from the `licence-updates` job we want to apply something to the workflow `data` property that will indicate why it was added. We could have just gone with `timeLimited: true` and that would have told us why it was added. But if we include the charge version ID instead it gives us something we can build on, for example, directly linking the manage workflow screen to the charge version with the time limited element.