Update process time limited service
We bring it in line with what we have done in our licence updates job.

For example, there is no need to check if the fetch service returned any results because we now process them in a loop (no results no looping).

We process them in a loop because it drops the dependence on the fetch service returning results that directly map to workflow records (the thing we broke and why we're making these changes). Plus it gives us more flexibility, for example, adding the charge version ID with the time limited charge element to the result.