Fix unique constraint issue in licences-update job
https://eaflood.atlassian.net/browse/WATER-4437

We encountered an issue when we first ran this job with real data. We should have spotted that the `water.charge_version_workflows` table has a unique constraint on its `licence_version_id` field.

```sql
CREATE UNIQUE INDEX unique_licence_version_id_date_deleted_null ON water.charge_version_workflows USING btree (licence_version_id) WHERE (date_deleted IS NULL);
```

This means a licence version won't have a `to_setup` workflow record but will have a `review` workflow record. Currently, we treat that as a 'result' so try to insert a new workflow record for the update licence.

This is causing an error. We're not familiar with charge version workflows enough to know whether removing the constraint would cause issues.

So, we just have to accept that an updated licence might appear to be in workflow, be dealt with, but then get put straight back in because now the blocking workflow record is gone so our job adds another back in for the licence.