Adding CSRF form protection via Hapi Crumb plugin
Crumb is used to diminish CSRF attacks using a random unique token that
is validated on the server side. Crumb may be used whenever you want to
prevent malicious code to execute system commands, that are performed by
HTTP requests.

This PR is focused on adding CSRF Tokens via the Hapi Crumb plugin.