Add calculated column to two-part tariff persisted data (#901)
https://eaflood.atlassian.net/browse/WATER-4370

As part of the current two-part tariff ticket we are working on, we need to display a banner on the match details page for when a charge elements allocated volume has been updated. The banner is always present once the allocated volume has been updated. To enable us to do this and to keep a record of the original calculated volume that the engine produced, we are adding a column to the persisted two-part tariff data called calculated volume. When the engine is run the first time both this column (calculated) and the allocated volume will have the same volume total persisted in it. If a user then updated the volume on the pages the new updated volume will be stored in allocated leaving calculated as the original calculated volume the engine produced. The banner can then be a simple comparison between these to fields. If they are the same, it hasn't been updated therefore no banner is needed and if they are difference then it ahs been updated and the banner is needed. Not only does this serve the purpose of the original ticket but it also means in the future if users want/need to view the calculated volume then we have that data available to show.