Fix 2PT volume allocation
https://eaflood.atlassian.net/browse/WATER-4377

It was found during testing that if the individual return line volumes are greater than the charge reference authorised volume, which in turn is greater than the element authorised volume. That the element was being over allocated.

This PR fixes this issue and creates a new unit test to check for it as it wasn't being picked up previously.