Fix fetching the returns for a two-part tariff licence
https://eaflood.atlassian.net/browse/WATER-4188

As part of the ongoing testing for the two-part tariff engine it was found that there were too many returns matching to a charge element. The returns were spread over two years (all within the financial year but starting and finishing outside). This wasn't a bug in the code but rather an A|C that wasn't fleshed out properly. The team don't want returns to be pulled in if the end date of the return is outside the financial year. This PR is to update the fetching query to adjust the start and end dates that the returns can be picked from.